fix: resolve 15 CodeQL scanning alerts in iccanalyzer-lite
- 11 narrow-loop-bound: cast uint8_t counters to int in comparisons
- 2 constant-comparison: use explicit unsigned/cast to avoid always-true
- 1 unused-local-variable: remove pcsSig (reserved for future use)
- 1 icc-buffer-overflow: use unsigned comparison for intent index
- 1 equality-on-floats: use relational ops instead of == for range check

254/254 tests pass.

+++ iccanalyzer-lite/IccHeuristicsTagValidation.cpp
@@ -118,7 +118,7 @@ printf("      Max safe CLUT entries per tag: %llu (16M)\n",
     icUInt8Number nIn = pMBB->InputChannels();
     uint64_t entries = 1;
     bool overflow = false;
-    for (int ch = 0; ch < nIn && ch < 16; ch++) {
+    for (int ch = 0; ch < static_cast<int>(nIn) && ch < 16; ch++) {
       if (!SafeMul64(&entries, entries, pCLUT->GridPoint(ch))) { overflow = true; break; }
     }
     if (!overflow) SafeMul64(&entries, entries, pCLUT->GetOutputChannels());
@@ -1142,7 +1142,7 @@ printf("[H28] LUT Dimension Validation (OOM Risk)\n");
           // Compute CLUT point count: nGrid^nInput * nOutput
           uint64_t points = 1;
           bool overflow28 = false;
-          for (int ch = 0; ch < nInput28; ch++) {
+          for (int ch = 0; ch < static_cast<int>(nInput28); ch++) {
             uint64_t prev = points;
             points *= nGrid28;
             if (nGrid28 > 0 && points / nGrid28 != prev) { overflow28 = true; break; }
